Whoops
This reverts commit 1b67c19da2637b5f0539025a6574a7a85c814f93.

+++ src/main/java/cloudburst/rejects/utils/WorldUtils.java
@@ -28,6 +28,6 @@ public class WorldUtils {
         double d = pos1.getX() - pos2.getX();
         double e = pos1.getY() - pos2.getY();
         double f = pos1.getZ() - pos2.getZ();
-        return MathHelper.sqrt((float) (d * d + e * e + f * f));
+        return MathHelper.sqrt(d * d + e * e + f * f);
     }
 }
